refactor: optimize scheduler and mooncake store code
