+++ lmdeploy/pytorch/kv_connector/mooncake/store/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .connector import MooncakeStoreConnector
 from .data import BlobBlockHashes, MooncakeStoreConfig, MooncakeStoreConnectorMetadata, MooncakeStoreRegistration
-from .worker import LookupKeyClient, LookupKeyServer
+from .lookup import LookupKeyClient, LookupKeyServer
 
 __all__ = [
     'BlobBlockHashes',
