+++ lmdeploy/pytorch/kv_connector/mooncake/store/worker.py
@@ -3,851 +3,44 @@
 
 from __future__ import annotations
 
-import fcntl
-import os
-import queue
-import socket
 import threading
 import time
-from collections import Counter
-from collections.abc import Callable, Mapping, Sequence
-from concurrent.futures import Future, ThreadPoolExecutor
-from dataclasses import dataclass
+from collections.abc import Mapping, Sequence
 from typing import TYPE_CHECKING, Any
-from uuid import uuid4
-
-import torch
-import zmq
 
 from lmdeploy.pytorch.kv_connector.base import KVCacheValue, KVConnectorOutput, RequestId
 from lmdeploy.utils import get_logger
 
 from .data import (
-    BlobBlockHashes,
     MooncakeStoreConfig,
     MooncakeStoreConnectorMetadata,
     MooncakeStoreKeyMetadata,
-    MooncakeStoreLoadRequest,
     MooncakeStoreRegistration,
-    MooncakeStoreSaveRequest,
     build_store_key,
     validate_kv_head_replica_num,
 )
-from .protocol import LOOKUP_MSG, RESET_MSG, RESP_ERR, RESP_OK
+from .lookup import LookupKeyClient, LookupKeyServer
+from .utils import (
+    StoreFactory,
+    _get_local_hostname,
+    _is_tensor,
+    _load_mooncake_store_factory,
+    _result_histogram,
+)
+from .worker_threads import KVCacheStoreRecvingThread, KVCacheStoreSendingThread
 
 if TYPE_CHECKING:
     from lmdeploy.pytorch.config import CacheConfig
 
 logger = get_logger('lmdeploy')
 
-StoreFactory = Callable[[], Any]
-
-_LOOKUP_POLL_INTERVAL_MS = 100
-_LOOKUP_RPC_TIMEOUT_MS = 5000
-
-
-def prepare_lookup_rpc_path(cache_config: CacheConfig) -> str:
-    """Create one endpoint before ``CacheConfig`` is copied to workers."""
-    transfer_config = cache_config.kv_transfer_config
-    if transfer_config is None or not transfer_config.is_kv_transfer_instance:
-        raise ValueError('lookup RPC requires an enabled kv_transfer_config')
-
-    extra_config = transfer_config.kv_connector_extra_config
-    configured_path = extra_config.get('lookup_rpc_path')
-    if configured_path is not None:
-        if not isinstance(configured_path, str) or not configured_path.startswith('ipc://'):
-            raise ValueError("lookup_rpc_path must be a non-empty 'ipc://' URI")
-        if configured_path == 'ipc://':
-            raise ValueError("lookup_rpc_path must be a non-empty 'ipc://' URI")
-        return configured_path
-
-    rpc_port = extra_config.get('lookup_rpc_port')
-    if rpc_port is None:
-        socket_path = f'ipc:///tmp/lmd-mc-lookup-{uuid4().hex}.sock'
-    else:
-        if isinstance(rpc_port, bool) or not isinstance(rpc_port, int) or rpc_port < 0:
-            raise ValueError('lookup_rpc_port must be a non-negative integer')
-        socket_path = f'ipc:///tmp/lmd-mc-lookup-{rpc_port}-{socket.gethostname()}.sock'
-    extra_config['lookup_rpc_path'] = socket_path
-    return socket_path
-
-
-def get_lookup_rpc_path(cache_config: CacheConfig) -> str:
-    """Return the endpoint shared by the scheduler and worker rank 0."""
-    return prepare_lookup_rpc_path(cache_config)
-
-
-def _get_lookup_rpc_timeout_ms(cache_config: CacheConfig) -> int:
-    transfer_config = cache_config.kv_transfer_config
-    assert transfer_config is not None
-    timeout_ms = transfer_config.kv_connector_extra_config.get('lookup_rpc_timeout_ms', _LOOKUP_RPC_TIMEOUT_MS)
-    if isinstance(timeout_ms, bool) or not isinstance(timeout_ms, int) or timeout_ms <= 0:
-        raise ValueError('lookup_rpc_timeout_ms must be a positive integer')
-    return timeout_ms
-
-
-def _make_zmq_socket(
-    context: zmq.Context,
-    socket_path: str,
-    socket_type: int,
-    *,
-    bind: bool,
-    timeout_ms: int | None = None,
-) -> zmq.Socket:
-    """Create one zero-linger lookup socket."""
-    rpc_socket = context.socket(socket_type)
-    rpc_socket.setsockopt(zmq.LINGER, 0)
-    if timeout_ms is not None:
-        rpc_socket.setsockopt(zmq.RCVTIMEO, timeout_ms)
-        rpc_socket.setsockopt(zmq.SNDTIMEO, timeout_ms)
-    try:
-        if bind:
-            rpc_socket.bind(socket_path)
-        else:
-            rpc_socket.connect(socket_path)
-    except Exception:
-        rpc_socket.close(linger=0)
-        raise
-    return rpc_socket
-
-
-def _load_mooncake_store_factory() -> StoreFactory:
-    """Import Mooncake only in a worker that actually enables the connector."""
-    try:
-        from mooncake.store import MooncakeDistributedStore
-    except ImportError as e:
-        raise ImportError(
-            'MooncakeStoreConnector requires the mooncake-transfer-engine package. '
-            'Install it before enabling the connector.') from e
-    return MooncakeDistributedStore
-
-
-def _load_mooncake_replicate_config() -> Any:
-    """Construct the default replication policy only when the first put
-    runs."""
-    try:
-        from mooncake.store import ReplicateConfig
-    except ImportError as e:
-        raise ImportError(
-            'Mooncake KV-cache saving requires ReplicateConfig from the '
-            'mooncake-transfer-engine package.') from e
-    return ReplicateConfig()
-
-
-def _get_local_hostname() -> str:
-    """Resolve the local address selected by the host routing table."""
-    candidates = (
-        (socket.AF_INET, ('8.8.8.8', 80)),
-        (socket.AF_INET6, ('2001:4860:4860::8888', 80)),
-    )
-    for family, remote_address in candidates:
-        try:
-            with socket.socket(family, socket.SOCK_DGRAM) as sock:
-                sock.connect(remote_address)
-                return str(sock.getsockname()[0])
-        except OSError:
-            continue
-    raise RuntimeError('cannot determine the local hostname for Mooncake Store')
-
-
-def _is_tensor(value: object) -> bool:
-    """Keep the production tensor check strict while allowing test patching."""
-    return isinstance(value, torch.Tensor)
-
-
-def _result_histogram(results: Sequence[int]) -> dict[int, int]:
-    return dict(sorted(Counter(results).items()))
-
-
-@dataclass
-class _StoreTask:
-    request: MooncakeStoreSaveRequest
-    ready_event: Any | None
-    enqueue_time: float
-    ready_waited: bool = False
-
-
-@dataclass
-class _LoadTask:
-    request: MooncakeStoreLoadRequest
-    enqueue_time: float
-
-
-class KVCacheStoreSendingThread(threading.Thread):
-    """Single-queue background writer for one Mooncake worker rank."""
-
-    _STOP = object()
-
-    def __init__(
-        self,
-        *,
-        store: Any,
-        registrations: tuple[MooncakeStoreRegistration, ...],
-        row_block_sizes: tuple[int, ...],
-        num_gpu_blocks: int,
-        key_metadata: MooncakeStoreKeyMetadata,
-        global_rank: int,
-        tp_rank: int,
-        tp_size: int,
-        completion_callback: Callable[[int], None],
-        replicate_config: Any | None = None,
-    ) -> None:
-        super().__init__(name='MooncakeKVCacheStoreSender', daemon=True)
-        if len(registrations) != len(row_block_sizes) or not registrations:
-            raise ValueError('Mooncake sender requires one block size per registered region')
-        if key_metadata.tp_size != tp_size:
-            raise ValueError(
-                'Mooncake sender tp_size must match key metadata: '
-                f'{tp_size} != {key_metadata.tp_size}')
-        if tp_rank < 0 or tp_rank >= tp_size:
-            raise ValueError(f'tp_rank must be in [0, {tp_size}), got {tp_rank}')
-        self.store = store
-        self.registrations = registrations
-        self.row_block_sizes = row_block_sizes
-        self.num_gpu_blocks = num_gpu_blocks
-        self.key_metadata = key_metadata
-        self.global_rank = global_rank
-        self.tp_rank = tp_rank
-        self.tp_size = tp_size
-        self.kv_head_replica_num = key_metadata.kv_head_replica_num
-        self.key_rank = tp_rank // self.kv_head_replica_num
-        self.replica_rank = tp_rank % self.kv_head_replica_num
-        self.completion_callback = completion_callback
-        self.replicate_config = replicate_config
-        self.request_queue: queue.Queue[_StoreTask | object] = queue.Queue()
-        self._state_lock = threading.Lock()
-        self._closed = False
-
-    def add_request(self, request: MooncakeStoreSaveRequest, ready_event: Any | None) -> None:
-        """Enqueue without waiting for GPU readiness or Store I/O."""
-        with self._state_lock:
-            if self._closed:
-                raise RuntimeError('Mooncake KV-cache sender is closed')
-            self.request_queue.put(_StoreTask(request, ready_event, time.perf_counter()))
-        logger.info(
-            'Mooncake KV save enqueued: global_rank=%d tp_rank=%d tp_size=%d '
-            'req_id=%s save_id=%d generation=%d token_len=%d blocks=%d',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-            request.generation,
-            request.token_len,
-            len(request.block_ids),
-        )
-
-    def _log_query_before(self, request: MooncakeStoreSaveRequest, keys: list[str]) -> float:
-        logger.info(
-            'Mooncake Store interaction before: operation=save_batch_is_exist '
-            'global_rank=%d tp_rank=%d tp_size=%d req_id=%s save_id=%d generation=%d '
-            'candidate_keys=%d first_key=%s last_key=%s',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-            request.generation,
-            len(keys),
-            keys[0],
-            keys[-1],
-        )
-        return time.perf_counter()
-
-    def _query_missing(self, request: MooncakeStoreSaveRequest, keys: list[str]) -> list[int] | None:
-        start = self._log_query_before(request, keys)
-        try:
-            exists_states = self.store.batch_is_exist(keys)
-            if not isinstance(exists_states, Sequence) or isinstance(exists_states, (str, bytes)):
-                raise TypeError('batch_is_exist must return a sequence')
-            exists_states = list(exists_states)
-            if len(exists_states) != len(keys):
-                raise ValueError(
-                    f'batch_is_exist returned {len(exists_states)} states for {len(keys)} keys')
-            if any(isinstance(state, bool) or not isinstance(state, int) or state not in (0, 1)
-                   for state in exists_states):
-                raise ValueError(f'batch_is_exist returned invalid states: {_result_histogram(exists_states)}')
-        except Exception as e:
-            logger.error(
-                'Mooncake Store interaction after: operation=save_batch_is_exist '
-                'global_rank=%d tp_rank=%d tp_size=%d req_id=%s save_id=%d generation=%d '
-                'status=error candidate_keys=%d elapsed_ms=%.3f error=%s',
-                self.global_rank,
-                self.tp_rank,
-                self.tp_size,
-                request.req_id,
-                request.save_id,
-                request.generation,
-                len(keys),
-                (time.perf_counter() - start) * 1000,
-                e,
-                exc_info=True,
-            )
-            return None
-
-        missing = [index for index, state in enumerate(exists_states) if state == 0]
-        logger.info(
-            'Mooncake Store interaction after: operation=save_batch_is_exist '
-            'global_rank=%d tp_rank=%d tp_size=%d req_id=%s save_id=%d generation=%d '
-            'status=ok candidate_keys=%d existing=%d missing=%d elapsed_ms=%.3f result_codes=%s',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-            request.generation,
-            len(keys),
-            len(keys) - len(missing),
-            len(missing),
-            (time.perf_counter() - start) * 1000,
-            _result_histogram(exists_states),
-        )
-        return missing
-
-    def _validate_block_id(self, block_id: int) -> None:
-        if isinstance(block_id, bool) or not isinstance(block_id, int):
-            raise TypeError('physical block IDs must be integers')
-        if block_id < 0 or block_id >= self.num_gpu_blocks:
-            raise ValueError(
-                f'physical block ID {block_id} is outside [0, {self.num_gpu_blocks})')
-
-    def _scatter_block(self, block_id: int) -> tuple[list[int], list[int]]:
-        self._validate_block_id(block_id)
-        addresses = [
-            registration.address + block_id * block_bytes
-            for registration, block_bytes in zip(
-                self.registrations,
-                self.row_block_sizes,
-                strict=True,
-            )
-        ]
-        return addresses, list(self.row_block_sizes)
-
-    def _wait_ready(self, task: _StoreTask) -> None:
-        if task.ready_waited:
-            return
-        task.ready_waited = True
-        if task.ready_event is None:
-            return
-        request = task.request
-        logger.info(
-            'Mooncake KV save wait before: global_rank=%d tp_rank=%d tp_size=%d '
-            'req_id=%s save_id=%d',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-        )
-        wait_start = time.perf_counter()
-        try:
-            task.ready_event.synchronize()
-        except Exception as e:
-            logger.error(
-                'Mooncake KV save wait after: global_rank=%d tp_rank=%d tp_size=%d '
-                'req_id=%s save_id=%d status=error elapsed_ms=%.3f error=%s',
-                self.global_rank,
-                self.tp_rank,
-                self.tp_size,
-                request.req_id,
-                request.save_id,
-                (time.perf_counter() - wait_start) * 1000,
-                e,
-                exc_info=True,
-            )
-            raise
-        logger.info(
-            'Mooncake KV save wait after: global_rank=%d tp_rank=%d tp_size=%d '
-            'req_id=%s save_id=%d status=ready elapsed_ms=%.3f',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-            (time.perf_counter() - wait_start) * 1000,
-        )
-
-    def _put_missing(
-        self,
-        task: _StoreTask,
-        keys: list[str],
-        missing_indices: list[int],
-        physical_block_ids: Sequence[int],
-    ) -> bool:
-        request = task.request
-        missing_keys = [keys[index] for index in missing_indices]
-        addresses = []
-        sizes = []
-        for index in missing_indices:
-            block_addresses, block_sizes = self._scatter_block(physical_block_ids[index])
-            addresses.append(block_addresses)
-            sizes.append(block_sizes)
-
-        total_bytes = sum(sum(block_sizes) for block_sizes in sizes)
-        logger.info(
-            'Mooncake Store interaction before: operation=save_batch_put_from_multi_buffers '
-            'global_rank=%d tp_rank=%d tp_size=%d req_id=%s save_id=%d generation=%d '
-            'keys=%d fragments_per_key=%d bytes=%d first_key=%s last_key=%s',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-            request.generation,
-            len(missing_keys),
-            len(self.registrations),
-            total_bytes,
-            missing_keys[0],
-            missing_keys[-1],
-        )
-        start = time.perf_counter()
-        try:
-            if self.replicate_config is None:
-                self.replicate_config = _load_mooncake_replicate_config()
-            results = self.store.batch_put_from_multi_buffers(
-                missing_keys,
-                addresses,
-                sizes,
-                self.replicate_config,
-            )
-            if not isinstance(results, Sequence) or isinstance(results, (str, bytes)):
-                raise TypeError('batch_put_from_multi_buffers must return a sequence')
-            results = list(results)
-            if len(results) != len(missing_keys):
-                raise ValueError(
-                    f'batch_put_from_multi_buffers returned {len(results)} results '
-                    f'for {len(missing_keys)} keys')
-            if any(isinstance(result, bool) or not isinstance(result, int) for result in results):
-                raise TypeError('batch_put_from_multi_buffers returned a non-integer result')
-        except Exception as e:
-            logger.error(
-                'Mooncake Store interaction after: operation=save_batch_put_from_multi_buffers '
-                'global_rank=%d tp_rank=%d tp_size=%d req_id=%s save_id=%d generation=%d '
-                'status=error keys=%d fragments_per_key=%d bytes=%d elapsed_ms=%.3f error=%s',
-                self.global_rank,
-                self.tp_rank,
-                self.tp_size,
-                request.req_id,
-                request.save_id,
-                request.generation,
-                len(missing_keys),
-                len(self.registrations),
-                total_bytes,
-                (time.perf_counter() - start) * 1000,
-                e,
-                exc_info=True,
-            )
-            return False
-
-        failed = sum(result < 0 for result in results)
-        status = 'ok' if failed == 0 else 'partial_failure'
-        log = logger.info if failed == 0 else logger.error
-        log(
-            'Mooncake Store interaction after: operation=save_batch_put_from_multi_buffers '
-            'global_rank=%d tp_rank=%d tp_size=%d req_id=%s save_id=%d generation=%d '
-            'status=%s keys=%d failed=%d fragments_per_key=%d bytes=%d elapsed_ms=%.3f '
-            'result_codes=%s',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.save_id,
-            request.generation,
-            status,
-            len(missing_keys),
-            failed,
-            len(self.registrations),
-            total_bytes,
-            (time.perf_counter() - start) * 1000,
-            _result_histogram(results),
-        )
-        return failed == 0
-
-    def _handle_request(self, task: _StoreTask) -> bool:
-        request = task.request
-        full_blocks = min(
-            request.token_len // self.key_metadata.block_size,
-            len(request.block_ids),
-            len(request.block_hashes),
-        )
-        if full_blocks == 0:
-            self._wait_ready(task)
-            return True
-
-        # Ranks holding an identical KV shard share one key namespace and
-        # stripe logical blocks across the replica group. Use the absolute
-        # request-local ordinal, not the physical cache block ID: physical IDs
-        # are allocator slots and can change when blocks are reused.
-        owned_ordinals = range(
-            self.replica_rank,
-            full_blocks,
-            self.kv_head_replica_num,
-        )
-        block_hashes = [request.block_hashes[index] for index in owned_ordinals]
-        block_ids = [request.block_ids[index] for index in owned_ordinals]
-        if not block_hashes:
-            # This save wave still owns the scheduler's block lease. Preserve
-            # the same forward-readiness fence before reporting completion even
-            # though this rank has no Store query or GPU read to perform.
-            self._wait_ready(task)
-            return True
-        for block_id in block_ids:
-            self._validate_block_id(block_id)
-        keys = [build_store_key(self.key_metadata, self.key_rank, block_hash) for block_hash in block_hashes]
-        missing_indices = self._query_missing(request, keys)
-        # Query does not touch GPU memory and can overlap the model write. The
-        # job still owns a read lease until this event, even when every key was
-        # already present or the query failed.
-        self._wait_ready(task)
-        if missing_indices is None:
-            return False
-        if not missing_indices:
-            return True
-        return self._put_missing(task, keys, missing_indices, block_ids)
-
-    def run(self) -> None:
-        while True:
-            item = self.request_queue.get()
-            try:
-                if item is self._STOP:
-                    return
-                assert isinstance(item, _StoreTask)
-                request = item.request
-                logger.info(
-                    'Mooncake KV save dequeued: global_rank=%d tp_rank=%d tp_size=%d '
-                    'req_id=%s save_id=%d generation=%d queue_wait_ms=%.3f',
-                    self.global_rank,
-                    self.tp_rank,
-                    self.tp_size,
-                    request.req_id,
-                    request.save_id,
-                    request.generation,
-                    (time.perf_counter() - item.enqueue_time) * 1000,
-                )
-                success = False
-                try:
-                    success = self._handle_request(item)
-                except Exception:
-                    logger.exception(
-                        'Mooncake KV save failed unexpectedly: global_rank=%d tp_rank=%d '
-                        'req_id=%s save_id=%d',
-                        self.global_rank,
-                        self.tp_rank,
-                        request.req_id,
-                        request.save_id,
-                    )
-                finally:
-                    if not item.ready_waited:
-                        try:
-                            self._wait_ready(item)
-                        except Exception:
-                            pass
-                    self.completion_callback(request.save_id)
-                    logger.info(
-                        'Mooncake KV save completed: global_rank=%d tp_rank=%d tp_size=%d '
-                        'req_id=%s save_id=%d generation=%d status=%s',
-                        self.global_rank,
-                        self.tp_rank,
-                        self.tp_size,
-                        request.req_id,
-                        request.save_id,
-                        request.generation,
-                        'ok' if success else 'error',
-                    )
-            finally:
-                self.request_queue.task_done()
-
-    def close(self) -> None:
-        """Stop accepting work, drain queued saves, and join exactly once."""
-        with self._state_lock:
-            if self._closed:
-                return
-            self._closed = True
-        self.request_queue.join()
-        self.request_queue.put(self._STOP)
-        self.request_queue.join()
-        self.join()
-
-
-class KVCacheStoreRecvingThread(threading.Thread):
-    """Single-queue background reader for one Mooncake worker rank.
-
-    The scheduler allocates and owns every destination block before a request
-    reaches this thread.  ``batch_get_into_multi_buffers`` therefore writes
-    directly into registered GPU cache rows without blocking model execution;
-    the request may resume only after the resulting load ID is completed by
-    every tensor-parallel worker.
-    """
-
-    _STOP = object()
-
-    def __init__(
-        self,
-        *,
-        store: Any,
-        registrations: tuple[MooncakeStoreRegistration, ...],
-        row_block_sizes: tuple[int, ...],
-        num_gpu_blocks: int,
-        key_metadata: MooncakeStoreKeyMetadata,
-        global_rank: int,
-        tp_rank: int,
-        tp_size: int,
-        completion_callback: Callable[[int, set[int]], None],
-    ) -> None:
-        super().__init__(name='MooncakeKVCacheStoreReceiver', daemon=True)
-        if len(registrations) != len(row_block_sizes) or not registrations:
-            raise ValueError('Mooncake receiver requires one block size per registered region')
-        if key_metadata.tp_size != tp_size:
-            raise ValueError(
-                'Mooncake receiver tp_size must match key metadata: '
-                f'{tp_size} != {key_metadata.tp_size}')
-        if tp_rank < 0 or tp_rank >= tp_size:
-            raise ValueError(f'tp_rank must be in [0, {tp_size}), got {tp_rank}')
-        self.store = store
-        self.registrations = registrations
-        self.row_block_sizes = row_block_sizes
-        self.num_gpu_blocks = num_gpu_blocks
-        self.key_metadata = key_metadata
-        self.global_rank = global_rank
-        self.tp_rank = tp_rank
-        self.tp_size = tp_size
-        self.key_rank = tp_rank // key_metadata.kv_head_replica_num
-        self.completion_callback = completion_callback
-        self.request_queue: queue.Queue[_LoadTask | object] = queue.Queue()
-        self._state_lock = threading.Lock()
-        self._closed = False
-
-    def add_request(self, request: MooncakeStoreLoadRequest) -> None:
-        """Enqueue a load without waiting for Store or GPU transfer work."""
-        with self._state_lock:
-            if self._closed:
-                raise RuntimeError('Mooncake KV-cache receiver is closed')
-            self.request_queue.put(_LoadTask(request, time.perf_counter()))
-        logger.info(
-            'Mooncake KV load enqueued: global_rank=%d tp_rank=%d tp_size=%d '
-            'req_id=%s load_id=%d generation=%d local_token_len=%d '
-            'remote_token_len=%d blocks=%d',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.load_id,
-            request.generation,
-            request.local_token_len,
-            request.remote_token_len,
-            len(request.block_ids),
-        )
-
-    def _validate_request(self, request: MooncakeStoreLoadRequest) -> None:
-        block_size = self.key_metadata.block_size
-        if request.local_token_len % block_size != 0:
-            raise ValueError('load local_token_len must be block aligned')
-        if request.remote_token_len % block_size != 0:
-            raise ValueError('load remote_token_len must be block aligned')
-        expected_blocks = (
-            request.remote_token_len - request.local_token_len
-        ) // block_size
-        if expected_blocks <= 0:
-            raise ValueError('load request must contain a non-empty external suffix')
-        if len(request.block_ids) != expected_blocks:
-            raise ValueError(
-                f'load request has {len(request.block_ids)} suffix blocks, '
-                f'expected {expected_blocks}')
-        if len(request.block_hashes) != expected_blocks:
-            raise ValueError(
-                f'load request has {len(request.block_hashes)} suffix hashes, '
-                f'expected {expected_blocks}')
-        for block_id in request.block_ids:
-            self._validate_block_id(block_id)
-
-    def _validate_block_id(self, block_id: int) -> None:
-        if isinstance(block_id, bool) or not isinstance(block_id, int):
-            raise TypeError('physical block IDs must be integers')
-        if block_id < 0 or block_id >= self.num_gpu_blocks:
-            raise ValueError(
-                f'physical block ID {block_id} is outside [0, {self.num_gpu_blocks})')
-
-    def _scatter_block(self, block_id: int) -> tuple[list[int], list[int]]:
-        self._validate_block_id(block_id)
-        addresses = [
-            registration.address + block_id * block_bytes
-            for registration, block_bytes in zip(
-                self.registrations,
-                self.row_block_sizes,
-                strict=True,
-            )
-        ]
-        return addresses, list(self.row_block_sizes)
-
-    def _load(self, request: MooncakeStoreLoadRequest) -> set[int]:
-        """Load the whole external suffix and return failed physical blocks."""
-        self._validate_request(request)
-
-        # KV-head replicas share one key namespace, but every TP rank needs a
-        # complete local GPU copy.  Unlike the save path, loads must never use
-        # replica-rank striding or block_id % replica_num filtering.
-        keys = [
-            build_store_key(self.key_metadata, self.key_rank, block_hash)
-            for block_hash in request.block_hashes
-        ]
-        addresses = []
-        sizes = []
-        for block_id in request.block_ids:
-            block_addresses, block_sizes = self._scatter_block(block_id)
-            addresses.append(block_addresses)
-            sizes.append(block_sizes)
-
-        total_bytes = sum(sum(block_sizes) for block_sizes in sizes)
-        logger.info(
-            'Mooncake Store interaction before: operation=load_batch_get_into_multi_buffers '
-            'global_rank=%d tp_rank=%d tp_size=%d req_id=%s load_id=%d generation=%d '
-            'local_token_len=%d remote_token_len=%d keys=%d fragments_per_key=%d '
-            'bytes=%d first_key=%s last_key=%s',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.load_id,
-            request.generation,
-            request.local_token_len,
-            request.remote_token_len,
-            len(keys),
-            len(self.registrations),
-            total_bytes,
-            keys[0],
-            keys[-1],
-        )
-        start = time.perf_counter()
-        try:
-            results = self.store.batch_get_into_multi_buffers(keys, addresses, sizes)
-            if not isinstance(results, Sequence) or isinstance(results, (str, bytes)):
-                raise TypeError('batch_get_into_multi_buffers must return a sequence')
-            results = list(results)
-            if len(results) != len(keys):
-                raise ValueError(
-                    f'batch_get_into_multi_buffers returned {len(results)} results '
-                    f'for {len(keys)} keys')
-            if any(isinstance(result, bool) or not isinstance(result, int) for result in results):
-                raise TypeError('batch_get_into_multi_buffers returned a non-integer result')
-        except Exception as e:
-            logger.error(
-                'Mooncake Store interaction after: operation=load_batch_get_into_multi_buffers '
-                'global_rank=%d tp_rank=%d tp_size=%d req_id=%s load_id=%d generation=%d '
-                'status=error keys=%d fragments_per_key=%d bytes=%d elapsed_ms=%.3f error=%s',
-                self.global_rank,
-                self.tp_rank,
-                self.tp_size,
-                request.req_id,
-                request.load_id,
-                request.generation,
-                len(keys),
-                len(self.registrations),
-                total_bytes,
-                (time.perf_counter() - start) * 1000,
-                e,
-                exc_info=True,
-            )
-            return set(request.block_ids)
-
-        failed_indices = [index for index, result in enumerate(results) if result < 0]
-        failed_block_ids = {request.block_ids[index] for index in failed_indices}
-        status = 'ok' if not failed_indices else 'partial_failure'
-        log = logger.info if not failed_indices else logger.error
-        log(
-            'Mooncake Store interaction after: operation=load_batch_get_into_multi_buffers '
-            'global_rank=%d tp_rank=%d tp_size=%d req_id=%s load_id=%d generation=%d '
-            'status=%s keys=%d failed=%d fragments_per_key=%d bytes=%d elapsed_ms=%.3f '
-            'result_codes=%s',
-            self.global_rank,
-            self.tp_rank,
-            self.tp_size,
-            request.req_id,
-            request.load_id,
-            request.generation,
-            status,
-            len(keys),
-            len(failed_indices),
-            len(self.registrations),
-            total_bytes,
-            (time.perf_counter() - start) * 1000,
-            _result_histogram(results),
-        )
-        return failed_block_ids
-
-    def run(self) -> None:
-        while True:
-            item = self.request_queue.get()
-            try:
-                if item is self._STOP:
-                    return
-                assert isinstance(item, _LoadTask)
-                request = item.request
-                logger.info(
-                    'Mooncake KV load dequeued: global_rank=%d tp_rank=%d tp_size=%d '
-                    'req_id=%s load_id=%d generation=%d queue_wait_ms=%.3f',
-                    self.global_rank,
-                    self.tp_rank,
-                    self.tp_size,
-                    request.req_id,
-                    request.load_id,
-                    request.generation,
-                    (time.perf_counter() - item.enqueue_time) * 1000,
-                )
-                failed_block_ids: set[int]
-                try:
-                    failed_block_ids = self._load(request)
-                except Exception:
-                    # Validation/address construction failures must still
-                    # complete the load so the scheduler can safely fall back
-                    # instead of leaving the request parked forever.
-                    failed_block_ids = set(request.block_ids)
-                    logger.exception(
-                        'Mooncake KV load failed unexpectedly: global_rank=%d tp_rank=%d '
-                        'req_id=%s load_id=%d',
-                        self.global_rank,
-                        self.tp_rank,
-                        request.req_id,
-                        request.load_id,
-                    )
-                try:
-                    self.completion_callback(request.load_id, failed_block_ids)
-                except Exception:
-                    # A bookkeeping callback must not kill the sole receiver
-                    # and strand all subsequently queued load waves.
-                    logger.exception(
-                        'Mooncake KV load completion callback failed: '
-                        'global_rank=%d tp_rank=%d req_id=%s load_id=%d',
-                        self.global_rank,
-                        self.tp_rank,
-                        request.req_id,
-                        request.load_id,
-                    )
-                logger.info(
-                    'Mooncake KV load completed: global_rank=%d tp_rank=%d tp_size=%d '
-                    'req_id=%s load_id=%d generation=%d status=%s failed_blocks=%d',
-                    self.global_rank,
-                    self.tp_rank,
-                    self.tp_size,
-                    request.req_id,
-                    request.load_id,
-                    request.generation,
-                    'ok' if not failed_block_ids else 'error',
-                    len(failed_block_ids),
-                )
-            finally:
-                self.request_queue.task_done()
-
-    def close(self) -> None:
-        """Stop accepting work, drain queued loads, and join exactly once."""
-        with self._state_lock:
-            if self._closed:
-                return
-            self._closed = True
-        self.request_queue.join()
-        self.request_queue.put(self._STOP)
-        self.request_queue.join()
-        self.join()
-
+__all__ = [
+    'KVCacheStoreRecvingThread',
+    'KVCacheStoreSendingThread',
+    'LookupKeyClient',
+    'LookupKeyServer',
+    'MooncakeStoreWorker',
+]
 
 class MooncakeStoreWorker:
     """Worker-side component of the Mooncake Store connector."""
@@ -924,7 +117,11 @@ class MooncakeStoreWorker:
 
         config_path = extra_config.get('mooncake_config_path')
         self.store_config = MooncakeStoreConfig.load_from_config(config_path)
-        local_hostname = _get_local_hostname()
+        local_hostname = extra_config.get('local_hostname')
+        if local_hostname is None:
+            local_hostname = _get_local_hostname()
+        elif not isinstance(local_hostname, str) or not local_hostname:
+            raise ValueError('local_hostname must be a non-empty string')
         factory = store_factory if store_factory is not None else _load_mooncake_store_factory()
         store = self._create_store(factory)
         try:
@@ -1578,335 +775,3 @@ class MooncakeStoreWorker:
             (time.perf_counter() - start) * 1000,
             ret,
         )
-
-
-class LookupKeyServer:
-    """ZMQ lookup server owned by Mooncake worker rank 0."""
-
-    def __init__(self, store_worker: MooncakeStoreWorker, cache_config: CacheConfig) -> None:
-        self.store_worker = store_worker
-        self.socket_path = get_lookup_rpc_path(cache_config)
-        self._ipc_path = self.socket_path.removeprefix('ipc://')
-        self._lock_path = f'{self._ipc_path}.lock'
-        self._lock_fd: int | None = None
-        self._owns_ipc_path = False
-        self._stop_event = threading.Event()
-        self._ready_event = threading.Event()
-        self._bind_error: BaseException | None = None
-        self._closed = False
-        self.running = True
-        self.context: zmq.Context | None = None
-        self.socket: zmq.Socket | None = None
-
-        self.thread = threading.Thread(
-            target=self._process_requests,
-            name='MooncakeLookupServer',
-            daemon=True,
-        )
-        self.thread.start()
-        if not self._ready_event.wait(timeout=5.0):
-            self._stop_event.set()
-            self.thread.join()
-            raise RuntimeError('LookupKeyServer did not start within 5 seconds')
-        if self._bind_error is not None:
-            self.thread.join()
-            raise RuntimeError(f'LookupKeyServer failed to bind {self.socket_path!r}') from self._bind_error
-
-    @staticmethod
-    def _frame_bytes(frame: Any) -> bytes:
-        return bytes(frame)
-
-    def _acquire_endpoint_lock(self) -> None:
-        lock_fd = os.open(self._lock_path, os.O_CREAT | os.O_RDWR, 0o600)
-        try:
-            fcntl.flock(lock_fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
-        except Exception:
-            os.close(lock_fd)
-            raise RuntimeError(f'lookup endpoint is already in use: {self.socket_path}') from None
-        self._lock_fd = lock_fd
-
-    def _release_endpoint_lock(self) -> None:
-        lock_fd = self._lock_fd
-        self._lock_fd = None
-        if lock_fd is None:
-            return
-        try:
-            if os.path.exists(self._lock_path):
-                os.unlink(self._lock_path)
-        finally:
-            fcntl.flock(lock_fd, fcntl.LOCK_UN)
-            os.close(lock_fd)
-
-    def _handle_lookup(self, frames: list[Any]) -> bytes:
-        if len(frames) != 4:
-            raise ValueError(f'lookup request must have 4 frames, got {len(frames)}')
-        token_frame = self._frame_bytes(frames[1])
-        hash_len_frame = self._frame_bytes(frames[2])
-        if len(token_frame) != 4 or len(hash_len_frame) != 2:
-            raise ValueError('lookup request has an invalid integer frame width')
-
-        token_len = int.from_bytes(token_frame, byteorder='big')
-        hash_len = int.from_bytes(hash_len_frame, byteorder='big')
-        payload_frame = frames[3]
-        if hasattr(payload_frame, 'buffer'):
-            payload = payload_frame.buffer
-        else:
-            payload = memoryview(self._frame_bytes(payload_frame))
-        block_hashes = BlobBlockHashes(payload, hash_len)
-        result = self.store_worker.lookup(token_len, block_hashes)
-        if (isinstance(result, bool) or not isinstance(result, int) or result < 0 or result > token_len
-                or result >= 2**32):
-            raise ValueError(f'lookup result must be a u32 integer, got {result!r}')
-        return result.to_bytes(4, byteorder='big')
-
-    def _handle_reset(self) -> bytes:
-        rank_fields_fn = getattr(self.store_worker, '_rank_fields', None)
-        rank_fields = rank_fields_fn() if rank_fields_fn is not None else (0, 0, 1)
-        logger.info(
-            'Mooncake Store interaction before: operation=remove_all global_rank=%d '
-            'tp_rank=%d tp_size=%d force=true',
-            *rank_fields,
-        )
-        start = time.perf_counter()
-        try:
-            recv_thread = getattr(self.store_worker, 'kv_recv_thread', None)
-            if recv_thread is not None:
-                recv_thread.request_queue.join()
-            send_thread = getattr(self.store_worker, 'kv_send_thread', None)
-            if send_thread is not None:
-                send_thread.request_queue.join()
-            store = self.store_worker.store
-            if store is None:
-                raise RuntimeError('MooncakeDistributedStore is not available')
-            store.remove_all(force=True)
-            logger.info(
-                'Mooncake Store interaction after: operation=remove_all global_rank=%d '
-                'tp_rank=%d tp_size=%d force=true status=ok elapsed_ms=%.3f',
-                *rank_fields,
-                (time.perf_counter() - start) * 1000,
-            )
-            return RESP_OK
-        except Exception as e:
-            logger.error(
-                'Mooncake Store interaction after: operation=remove_all global_rank=%d '
-                'tp_rank=%d tp_size=%d force=true status=error elapsed_ms=%.3f error=%s',
-                *rank_fields,
-                (time.perf_counter() - start) * 1000,
-                e,
-                exc_info=True,
-            )
-            return RESP_ERR
-
-    def _dispatch(self, frames: list[Any]) -> bytes:
-        if not frames:
-            logger.warning('LookupKeyServer received an empty request.')
-            return RESP_ERR
-        msg_type = self._frame_bytes(frames[0])
-        if msg_type == LOOKUP_MSG:
-            try:
-                return self._handle_lookup(frames)
-            except Exception as e:
-                logger.error('Mooncake lookup request failed: %s', e, exc_info=True)
-                return (0).to_bytes(4, byteorder='big')
-        if msg_type == RESET_MSG:
-            if len(frames) != 1:
-                logger.warning('LookupKeyServer received a malformed reset request.')
-                return RESP_ERR
-            return self._handle_reset()
-        logger.warning('LookupKeyServer received unknown msg_type: %r', msg_type)
-        return RESP_ERR
-
-    def _process_requests(self) -> None:
-        try:
-            self.context = zmq.Context()
-            self._acquire_endpoint_lock()
-            if os.path.exists(self._ipc_path):
-                os.unlink(self._ipc_path)
-            self.socket = _make_zmq_socket(self.context, self.socket_path, zmq.REP, bind=True)
-            self._owns_ipc_path = True
-            self._ready_event.set()
-            while not self._stop_event.is_set():
-                assert self.socket is not None
-                if self.socket.poll(_LOOKUP_POLL_INTERVAL_MS, zmq.POLLIN) == 0:
-                    continue
-                frames = self.socket.recv_multipart(copy=False)
-                self.socket.send(self._dispatch(frames))
-        except Exception as e:
-            if not self._ready_event.is_set():
-                self._bind_error = e
-                self._ready_event.set()
-                return
-            if isinstance(e, zmq.ContextTerminated):
-                return
-            if isinstance(e, zmq.ZMQError) and self._stop_event.is_set():
-                return
-            logger.exception('LookupKeyServer request loop failed.')
-        finally:
-            if self.socket is not None:
-                self.socket.close(linger=0)
-                self.socket = None
-            if self.context is not None:
-                self.context.term()
-                self.context = None
-            if self._owns_ipc_path and os.path.exists(self._ipc_path):
-                os.unlink(self._ipc_path)
-            self._owns_ipc_path = False
-            self._release_endpoint_lock()
-            self.running = False
-
-    def close(self) -> None:
-        """Stop the server thread and remove its IPC socket path."""
-        if self._closed:
-            return
-        self._closed = True
-        self._stop_event.set()
-        self.thread.join()
-
-
-class LookupKeyClient:
-    """Scheduler-side ZMQ client with non-blocking Future polling."""
-
-    def __init__(self, cache_config: CacheConfig) -> None:
-        self.cache_config = cache_config
-        self.socket_path = get_lookup_rpc_path(cache_config)
-        self.timeout_ms = _get_lookup_rpc_timeout_ms(cache_config)
-        self.context: zmq.Context | None = None
-        self.socket: zmq.Socket | None = None
-        self.executor = ThreadPoolExecutor(max_workers=1, thread_name_prefix='MooncakeLookupClient')
-        self.futures: dict[RequestId, Future[int]] = {}
-        self._closed = False
-
-    def _make_socket(self) -> zmq.Socket:
-        assert self.context is not None
-        return _make_zmq_socket(
-            self.context,
-            self.socket_path,
-            zmq.REQ,
-            bind=False,
-            timeout_ms=self.timeout_ms,
-        )
-
-    def _ensure_socket(self) -> zmq.Socket:
-        if self.socket is None:
-            if self.context is None:
-                self.context = zmq.Context()
-            self.socket = self._make_socket()
-        return self.socket
-
-    def _reconnect_socket(self) -> None:
-        if self.socket is not None:
-            self.socket.close(linger=0)
-            self.socket = None
-        if not self._closed:
-            self.socket = self._make_socket()
-
-    def _close_transport(self) -> None:
-        if self.socket is not None:
-            self.socket.close(linger=0)
-            self.socket = None
-        if self.context is not None:
-            self.context.term()
-            self.context = None
-
-    @staticmethod
-    def _encode_lookup(token_len: int, block_hashes: Sequence[bytes]) -> tuple[bytes, bytes, bytes, bytes]:
-        if isinstance(token_len, bool) or not isinstance(token_len, int) or token_len < 0 or token_len >= 2**32:
-            raise ValueError('token_len must be a u32 integer')
-
-        hashes = []
-        for block_hash in block_hashes:
-            if not isinstance(block_hash, (bytes, bytearray, memoryview)):
-                raise TypeError('block hashes must be bytes-like values')
-            hashes.append(bytes(block_hash))
-        hash_len = len(hashes[0]) if hashes else 0
-        if hashes and hash_len == 0:
-            raise ValueError('block hashes must not be empty')
-        if hash_len >= 2**16:
-            raise ValueError('block hash length must fit in a u16 integer')
-        if any(len(block_hash) != hash_len for block_hash in hashes):
-            raise ValueError('all block hashes must have the same length')
-
-        return (
-            LOOKUP_MSG,
-            token_len.to_bytes(4, byteorder='big'),
-            hash_len.to_bytes(2, byteorder='big'),
-            b''.join(hashes),
-        )
-
-    def _lookup(self, token_len: int, block_hashes: Sequence[bytes]) -> int:
-        frames = self._encode_lookup(token_len, block_hashes)
-        rpc_socket = self._ensure_socket()
-        try:
-            rpc_socket.send_multipart(frames, copy=False)
-            response = rpc_socket.recv()
-        except zmq.ZMQError:
-            self._reconnect_socket()
-            raise
-        if len(response) != 4:
-            raise RuntimeError(f'lookup response must have 4 bytes, got {len(response)}')
-        result = int.from_bytes(response, byteorder='big')
-        if result > token_len:
-            raise RuntimeError(f'lookup response {result} exceeds token_len {token_len}')
-        return result
-
-    def lookup(
-        self,
-        req_id: RequestId,
-        token_len: int,
-        block_hashes: Sequence[bytes],
-        non_block: bool = True,
-    ) -> int | None:
-        """Submit once per request and poll without blocking by default."""
-        if self._closed:
-            raise RuntimeError('LookupKeyClient is closed')
-        future = self.futures.get(req_id)
-        if future is None:
-            future = self.executor.submit(self._lookup, token_len, tuple(block_hashes))
-            self.futures[req_id] = future
-        if non_block and not future.done():
-            return None
-        try:
-            return future.result()
-        except Exception as e:
-            logger.error('Asynchronous Mooncake lookup failed for %s: %s', req_id, e, exc_info=True)
-            return 0
-        finally:
-            self.futures.pop(req_id, None)
-
-    def discard(self, req_id: RequestId) -> None:
-        """Drop an in-flight or completed lookup for an aborted request."""
-        future = self.futures.pop(req_id, None)
-        if future is not None:
-            future.cancel()
-
-    def _reset(self) -> bool:
-        rpc_socket = self._ensure_socket()
-        try:
-            rpc_socket.send(RESET_MSG)
-            response = rpc_socket.recv()
-        except zmq.ZMQError:
-            self._reconnect_socket()
-            raise
-        return bytes(response) == RESP_OK
-
-    def reset(self) -> bool:
-        if self._closed:
-            raise RuntimeError('LookupKeyClient is closed')
-        try:
-            return self.executor.submit(self._reset).result()
-        except Exception as e:
-            logger.error('Mooncake reset RPC failed: %s', e, exc_info=True)
-            return False
-
-    def close(self) -> None:
-        """Cancel queued lookups and release the client socket exactly once."""
-        if self._closed:
-            return
-        self._closed = True
-        futures = list(self.futures.values())
-        self.futures.clear()
-        for future in futures:
-            future.cancel()
-        close_future = self.executor.submit(self._close_transport)
-        close_future.result()
-        self.executor.shutdown(wait=True, cancel_futures=True)
